requirements/base.txt
# Django core
Django==5.2.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-environ==0.12.0
django-cors-headers==4.8.0
drf-spectacular==0.28.0

whitenoise==6.10.0

# Postgres
psycopg2-binary==2.9.10

# Utils
requests==2.32.5
inflection==0.5.1
python-dateutil==2.9.0.post0
tzdata==2025.2
urllib3==2.5.0
idna==3.10
charset-normalizer==3.4.3
certifi==2025.8.3
sqlparse==0.5.3
typing_extensions==4.15.0
six==1.17.0
asgiref==3.9.1
